Fixed bug in k nearest neighbors.
Found a bug in insert_min function, where it was possible
for a point inside the current search radius to replace
a point in the mins_so_far list that was closer to the
query point.

Fixed this by checking the distance of the pointed to be
inserted against the actual maximum distance in the
list, instead of just using mins_so_far['max_distance'],
which is technically the search radius and not the
distance of the fartherest point in the mins_so_far
list.